+++ src/services/storage.js
@@ -10,3 +10,13 @@ export const saveToSettings = value => {
   const valueString = JSON.stringify(value);
   return AsyncStorage.setItem("@Settings:Queue", valueString);
 };
+
+export const saveConfig = settings => {
+  return saveToSettings({
+    deviceId: settings.deviceId,
+    queueHost: settings.queueHost,
+    hasTailOrder: settings.hasTailOrder,
+    useDescription: settings.useDescription,
+    useDefaultCustomer: settings.useDefaultCustomer,
+  });
+};