+++ ChangeLog.md
@@ -55,3 +55,4 @@
 
 # v1.2.7 [10 мая 2019]
 - Fix on protocol in syncing
+- Added Device Id field